requirements.txt
bcrypt==3.2.0
celery==5.1.2
flask==2.0.2
flask-session==0.4.0
numpy==1.19.5
peewee==3.14.8
Pillow==8.1.0
requests==2.24.0